Change config 'input_col' to 'substance_input_col' for clarity
This change is intended to clarify that the user should input the OSHA
substance name to CompTox rather than other possible identifiers, such
as DTXSID, CASRN, etc. This requirement is made because the CompTox
input column is used to map the OSHA substances column directly to
parameters such as molecular weight.